DOMLauncher .9 Beta!
Tutorial theme is finished.  Css is generated for each of the dpi's.
Swapped a variable name for the Screenbrightness plugin.  There is no
'percentage'.  Volume controls got expanded to return a percentage as an
option when checking the audio value.  Theme is imageless now.  CSS/JS
should be commented out.